# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecommendationInput": {
        "additionalProperties": false,
        "properties": {
          "expected_history_version": {
            "minimum": 0.0,
            "title": "Expected History Version",
            "type": "integer"
          },
          "k": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "schema_version": {
            "const": "0.1",
            "default": "0.1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "strategy": {
            "$ref": "#/components/schemas/Strategy",
            "default": "popular"
          }
        },
        "required": [
          "session_id",
          "expected_history_version"
        ],
        "title": "RecommendationInput",
        "type": "object"
      },
      "RecommendationItemResponse": {
        "properties": {
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "score",
          "source"
        ],
        "title": "RecommendationItemResponse",
        "type": "object"
      },
      "RecommendationResponse": {
        "properties": {
          "actual_strategy": {
            "title": "Actual Strategy",
            "type": "string"
          },
          "bundle_id": {
            "format": "uuid",
            "title": "Bundle Id",
            "type": "string"
          },
          "captured_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At Ms"
          },
          "exclusion_version": {
            "title": "Exclusion Version",
            "type": "integer"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "history_version": {
            "title": "History Version",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "requested_strategy": {
            "title": "Requested Strategy",
            "type": "string"
          },
          "session_epoch": {
            "title": "Session Epoch",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "session_id",
          "session_epoch",
          "history_version",
          "bundle_id",
          "exclusion_version",
          "requested_strategy",
          "actual_strategy",
          "fallback_reason",
          "items"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
      "SessionCreateInput": {
        "additionalProperties": false,
        "properties": {
          "profile_id": {
            "default": "new",
            "enum": [
              "new",
              "sample"
            ],
            "title": "Profile Id",
            "type": "string"
          }
        },
        "title": "SessionCreateInput",
        "type": "object"
      },
      "SessionCreatedResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "epoch": {
            "title": "Epoch",
            "type": "integer"
          },
          "favorite_items": {
            "items": {
              "type": "string"
            },
            "title": "Favorite Items",
            "type": "array"
          },
          "hidden_items": {
            "items": {
              "type": "string"
            },
            "title": "Hidden Items",
            "type": "array"
          },
          "history": {
            "items": {
              "type": "string"
            },
            "title": "History",
            "type": "array"
          },
          "history_version": {
            "title": "History Version",
            "type": "integer"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "epoch",
          "history_version",
          "history",
          "hidden_items",
          "favorite_items",
          "profile_id",
          "access_token"
        ],
        "title": "SessionCreatedResponse",
        "type": "object"
      },
...
      "Strategy": {
        "enum": [
          "popular",
          "dense",
          "generative",
          "hybrid",
          "adaptive"
        ],
        "title": "Strategy",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "本地推荐与管理演示。未配置 PostgreSQL：会话和推荐结果只保存在当前进程。受控 CPU bundle 可发布；R06 冻结语料服务须显式启用并批准发布，不支持任意新商品。",
    "title": "EvoRec Demo API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/recommendations": {
      "post": {
        "operationId": "recommend_api_v1_recommendations_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid session token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Session not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "History, idempotency or execution recovery conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "R06 CPU capacity exhausted"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Catalog, runtime or snapshot unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recommendation deadline exceeded"
          }
        },
        "summary": "Recommend",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/v1/sessions": {
      "post": {
        "operationId": "create_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SessionCreateInput"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Fixed sample is unavailable in the active catalog"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}